sync/requirements.txt
psycopg2-binary>=2.9.0
firebase-admin>=6.0.0
schedule>=1.2.0
